Removed AddThreadSelectorEventLoop
